# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/login":{"post":{"tags":["auth"],"summary":"Login For Access Token","operationId":"login_for_access_token_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_for_access_token_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/register":{"post":{"tags":["auth"],"summary":"Register","operationId":"register_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/me":{"get":{"tags":["auth"],"summary":"Read Users Me","operationId":"read_users_me_users_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserBase"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/logout":{"post":{"tags":["auth"],"summary":"Logout","operationId":"logout_logout_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/years":{"get":{"tags":["entries"],"summary":"Read Years","operationId":"read_years_years_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"post":{"tags":["entries"],"summary":"Create Year","operationId":"create_year_years_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/YearCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/years/{year_id}":{"delete":{"tags":["entries"],"summary":"Delete Year","operationId":"delete_year_years__year_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"year_id","in":"path","required":true,"schema":{"type":"integer","title":"Year Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/years/{year_id}/entries":{"get":{"tags":["entries"],"summary":"Read Entries","operationId":"read_entries_years__year_id__entries_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"year_id","in":"path","required":true,"schema":{"type":"integer","title":"Year Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["entries"],"summary":"Create Revenue","operationId":"create_revenue_years__year_id__entries_post","security":[{"OAuth2PasswordBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EntriesCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/years/{year_id}/entries/{entry_id}":{"delete":{"tags":["entries"],"summary":"Delete Entry","operationId":"delete_entry_years__year_id__entries__entry_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"year_id","in":"path","required":true,"schema":{"type":"integer","title":"Year Id"}},{"name":"entry_id","in":"path","required":true,"schema":{"type":"integer","title":"Entry Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["entries"],"summary":"Update Entry","operationId":"update_entry_years__year_id__entries__entry_id__put","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"year_id","in":"path","required":true,"schema":{"type":"integer","title":"Year Id"}},{"name":"entry_id","in":"path","required":true,"schema":{"type":"integer","title":"Entry Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EntriesCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["entries"],"summary":"Read Entry","operationId":"read_entry_years__year_id__entries__entry_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"year_id","in":"path","required":true,"schema":{"type":"integer","title":"Year Id"}},{"name":"entry_id","in":"path","required":true,"schema":{"type":"integer","title":"Entry Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/years/{year_id}/profit":{"get":{"tags":["entries"],"summary":"Read Profit","operationId":"read_profit_years__year_id__profit_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"year_id","in":"path","required":true,"schema":{"type":"integer","title":"Year Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/years/uploads/{file_id}":{"get":{"tags":["entries"],"summary":"Read File","operationId":"read_file_years_uploads__file_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"file_id","in":"path","required":true,"schema":{"type":"integer","title":"File Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/years/uploads":{"post":{"tags":["entries"],"summary":"Upload File","operationId":"upload_file_years_uploads_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_file_years_uploads_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/years/{year_id}/export/csv":{"get":{"tags":["entries"],"summary":"Export Csv","operationId":"export_csv_years__year_id__export_csv_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"year_id","in":"path","required":true,"schema":{"type":"integer","title":"Year Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/years/{year_id}/export/pdf":{"get":{"tags":["entries"],"summary":"Export Pdf","operationId":"export_pdf_years__year_id__export_pdf_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"year_id","in":"path","required":true,"schema":{"type":"integer","title":"Year Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/categories":{"get":{"tags":["entries"],"summary":"Read Categories","operationId":"read_categories_categories_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}}},"components":{"schemas":{"Body_login_for_access_token_login_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"password"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_for_access_token_login_post"},"Body_upload_file_years_uploads_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_file_years_uploads_post"},"EntriesCreate":{"properties":{"revenue":{"type":"number","title":"Revenue"},"cost":{"type":"number","title":"Cost"},"date":{"type":"integer","title":"Date"},"file_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"File Id"},"year_id":{"type":"integer","title":"Year Id"},"cat_id":{"type":"integer","title":"Cat Id"}},"type":"object","required":["revenue","cost","date","year_id","cat_id"],"title":"EntriesCreate"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type"}},"type":"object","required":["access_token","token_type"],"title":"Token"},"UserBase":{"properties":{"username":{"type":"string","title":"Username"}},"type":"object","required":["username"],"title":"UserBase"},"UserCreate":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","password"],"title":"UserCreate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"YearCreate":{"properties":{"year":{"type":"integer","maximum":2024.0,"minimum":2010.0,"title":"Year"}},"type":"object","required":["year"],"title":"YearCreate"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"login"}}}}}}